v0.44

+++ src/com/github/kuben/realshopping/ShippedPackage.java
@@ -1,3 +1,22 @@
+/*
+ * RealShopping Bukkit plugin for Minecraft
+ * Copyright 2013 Jakub Fojt
+ * 
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * 
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *     
+ */
+
 package com.github.kuben.realshopping;
 
 import java.util.Date;
@@ -9,18 +28,18 @@ import org.bukkit.inventory.ItemStack;
 public class ShippedPackage{
 	
 	private ItemStack[] contents;
-	private float cost;
+	private int cost;
 	private long dateSent;
 	private Location locationSent;
 	
-	public ShippedPackage(ItemStack[] contents, float cost, Location locationSent){
+	public ShippedPackage(ItemStack[] contents, int cost, Location locationSent){
 		this.contents = contents;
 		this.cost = cost;
 		this.locationSent = locationSent;
 		this.dateSent = System.currentTimeMillis();
 	}
 	
-	public ShippedPackage(ItemStack[] contents, float cost, Location locationSent, long dateSent){
+	public ShippedPackage(ItemStack[] contents, int cost, Location locationSent, long dateSent){
 		this.contents = contents;
 		this.cost = cost;
 		this.locationSent = locationSent;
@@ -35,7 +54,7 @@ public class ShippedPackage{
 		return dateSent;
 	}
 	
-	public float getCost(){
+	public int getCost(){
 		return cost;
 	}
